NAV-28505 Ordnet for en tredje tilstand som er hvis man får på Gjenbruk, men ikke har fylt ut Dokument egenskaper + gjenbruk
