+++ apps/frontend-nextjs/app/components/etterlevelseDokumentasjon/etterlevelseDokumentasjonPage/gjenbruk/TillatGjenbrukModal.tsx
@@ -127,6 +127,7 @@ export const TillatGjenbrukModal: FunctionComponent<TProps> = ({
                       isSubmitting={isSubmitting}
                       hasMissingRequiredField={hasMissingRequiredField}
                       setSubmitClick={setSubmitClick}
+                      submit={submit}
                       initialValues={initialValues}
                     />
                   )}
