--- apps/frontend-nextjs/app/components/etterlevelseDokumentasjon/etterlevelseDokumentasjonPage/gjenbruk/common/avbrytKnapp.tsx
@@ -1,13 +1,0 @@
-import { Button } from '@navikt/ds-react'
-import { FunctionComponent } from 'react'
-
-type TProps = {
-  isSubmitting: boolean
-  setIsOpen: (open: boolean) => void
-}
-
-export const AvbrytKnapp: FunctionComponent<TProps> = ({ isSubmitting, setIsOpen }) => (
-  <Button type='button' disabled={isSubmitting} variant='tertiary' onClick={() => setIsOpen(false)}>
-    Avbryt
-  </Button>
-)
