+++ apps/frontend-nextjs/app/components/etterlevelseDokumentasjon/etterlevelseDokumentasjonPage/gjenbruk/tilgjengeligForGjenbrukModal/tilgjengeligForGjenbrukModal.tsx
@@ -3,7 +3,7 @@ import { Button, InlineMessage, Modal } from '@navikt/ds-react'
 import { FormikErrors } from 'formik'
 import { Dispatch, FunctionComponent, SetStateAction } from 'react'
 import { BeskrivelseAvGjenbruk } from '../beskrivelseAvGjenbruk/beskrivelseAvGjenbruk'
-import { AvbrytKnapp } from '../common/avbrytKnapp'
+import { AvbrytKnapp, LagreTilSenereKnapp } from '../common/knapper'
 import { GjenbrukFeilmelding } from '../feilmelding/feilmelding'
 
 type TProps = {
@@ -19,6 +19,7 @@ type TProps = {
   hasMissingRequiredField: boolean
   setSubmitClick: Dispatch<SetStateAction<boolean>>
   submitForm: (() => Promise<void>) & (() => Promise<any>)
+  submit: (etterlevelseDokumentasjon: TEtterlevelseDokumentasjonQL) => Promise<void>
   initialValues: TEtterlevelseDokumentasjonQL
 }
 
@@ -31,6 +32,7 @@ export const TilgjengeligForGjenbrukModal: FunctionComponent<TProps> = ({
   hasMissingRequiredField,
   setSubmitClick,
   submitForm,
+  submit,
   initialValues,
 }) => (
   <>
@@ -51,7 +53,6 @@ export const TilgjengeligForGjenbrukModal: FunctionComponent<TProps> = ({
         variant='primary'
         disabled={isSubmitting || hasMissingRequiredField}
         onClick={async () => {
-          if (hasMissingRequiredField) return
           setSubmitClick((prev) => !prev)
           await setFieldValue('tilgjengeligForGjenbruk', true)
           await submitForm()
@@ -64,7 +65,7 @@ export const TilgjengeligForGjenbrukModal: FunctionComponent<TProps> = ({
         <Button
           type='button'
           variant='secondary'
-          disabled={isSubmitting}
+          disabled={isSubmitting || hasMissingRequiredField}
           onClick={async () => {
             setSubmitClick((prev) => !prev)
             await setFieldValue('tilgjengeligForGjenbruk', false)
@@ -73,6 +74,24 @@ export const TilgjengeligForGjenbrukModal: FunctionComponent<TProps> = ({
         >
           Slå av gjenbruk
         </Button>
+      )}
+
+      {initialValues.tilgjengeligForGjenbruk && initialValues.gjenbrukBeskrivelse.length === 0 && (
+        <LagreTilSenereKnapp
+          isSubmitting={isSubmitting}
+          setSubmitClick={setSubmitClick}
+          submit={submit}
+          values={values}
+        />
+      )}
+
+      {initialValues.tilgjengeligForGjenbruk && initialValues.gjenbrukBeskrivelse.length > 0 && (
+        <LagreTilSenereKnapp
+          isSubmitting={isSubmitting}
+          setSubmitClick={setSubmitClick}
+          submit={submit}
+          values={values}
+        />
       )}
 
       <AvbrytKnapp isSubmitting={isSubmitting} setIsOpen={setIsOpen} />
